[wireless-regdb] Install regdb to libdir. JB#48318
There are adaptations which have /lib/firmware as a symlink. To avoid
conflicts of packaging or writing the regdb files to wrong location
through the symlink modify the firmware install location in this
package. If some adaptation really needs the files to reside in
/lib/firmware they can be symlinked per adaptation.

If need be the firmware files can be moved back to /lib/firmware when
next stop release is made.